+++ grey/x.cpp
@@ -190,7 +190,7 @@ EXPORTED void tooltip(const char* text) {
     w::tooltip(text);
 }
 
-EXPORTED bool combo(const char* c_label, const char** options, int32_t options_size, size_t* selected, float width) {
+EXPORTED bool combo(const char* c_label, const char** options, int32_t options_size, uint32_t* selected, float width) {
     return w::combo(c_label, vector<string>(options, options + options_size), *selected, width * scale);
 }
